image_split_kmeans/requirements.txt
opencv-python
numpy
scikit-learn
matplotlib
Pillow
pyyaml
